+++ packages/eslint-plugin/rules/dependency-group.js
@@ -6,7 +6,21 @@ module.exports = {
 			url:
 				'https://github.com/WordPress/gutenberg/blob/master/packages/eslint-plugin/docs/rules/dependency-group.md',
 		},
-		schema: [],
+		schema: [
+			{
+				type: 'object',
+				properties: {
+					isInternalDependencyPatterns: {
+						type: 'array',
+						items: {
+							type: 'string',
+						},
+						uniqueItems: true,
+					},
+				},
+				additionalProperties: false,
+			},
+		],
 		fixable: true,
 		messages: {
 			expectExternal:
@@ -155,6 +169,9 @@ module.exports = {
 
 		return {
 			Program( node ) {
+				const options = context.options[ 0 ] || {};
+				const { isInternalDependencyPatterns = [] } = options;
+
 				/**
 				 * The set of package localities which have been reported for
 				 * the current program. Each locality is reported at most one
@@ -192,7 +209,16 @@ module.exports = {
 						return;
 					}
 
-					const locality = getPackageLocality( source );
+					let locality = getPackageLocality( source );
+
+					const isInternalDependency = isInternalDependencyPatterns.some(
+						( pattern ) => source.match( pattern )
+					);
+
+					if ( isInternalDependency ) {
+						locality = 'Internal';
+					}
+
 					if ( verified.has( locality ) ) {
 						return;
 					}